adding agency modal
