+++ .idea/jsLibraryMappings.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="JavaScriptLibraryMappings">
-    <file url="file://$PROJECT_DIR$" libraries="{express_app0/node_modules}" />
+    <file url="PROJECT" libraries="{bootstrap, jquery}" />
     <includedPredefinedLibrary name="Node.js Core" />
   </component>
 </project>
