+++ lib/database.js
@@ -1,0 +1,16 @@
+/**
+ * Created by Tony on 2/23/2017.
+ */
+
+var mysql     =    require('mysql');
+
+var pool      =    mysql.createPool({
+    connectionLimit : 100, //important
+    host     : 'localhost',
+    user     : 'root',
+    password : 'Burnstuff1302',
+    database : 'mlk_day',
+    debug    :  false
+});
+
+module.exports = pool;
