+++ views/Header.ejs
@@ -1,1 +1,6 @@
-<h1> MLK Day Volunteering</h1>
+
+    <div >
+        <div class="container">
+            <img class ="banner" src ="images/MLK17.png" alt = "MLK banner image" width = "100%"/>
+        </div>
+    </div>
