+++ views/admin.ejs
@@ -1,12 +1,68 @@
+<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
 <%include Header%>
 
 <h1>Welcome to the administration page</h1>
-<form method="post" action="/admin">
-<label>Profile Search</label>
-<input type="email" name="search">
-    <input type="submit" value="search">
+<form method = "post" class="form-horizontal" role="form" action = "/admin">
+    <div class="form-group">
+        <label for="email" class="col-sm-3 control-label">Email</label>
+        <div class="col-sm-9">
+            <input type="email" name="Email" id="email" placeholder="Email" class="form-control">
+        </div>
+        <div class="form-group">
+            <div class="col-sm-9 col-sm-offset-3">
+                <button type="submit" name ="search" class="btn btn-primary btn-block" value="search">Search</button>
+            </div>
+        </div>
+        <div class="col-sm-9 col-sm-offset-3">
+            <button type="button" class="btn btn-primary btn-block" data-toggle="modal" data-target="#myModal">Add Agency</button>
+        </div>
+    </div>
 </form>
 
+</form> <!-- /form -->
 <form method="post" action="/admin">
-    <input type="submit" name="dump" value="Data Dump">
+    <div class="form-group">
+        <div class="col-sm-9 col-sm-offset-3">
+            <button type="submit" class="btn btn-primary btn-block" value="Data Dump">Dump</button>
+        </div>
+    </div>
 </form>
+
+
+<!-- Modal -->
+<div id="myModal" class="modal fade" role="dialog">
+    <div class="modal-dialog">
+
+        <!-- Modal content-->
+        <div class="modal-content">
+            <div class="modal-header">
+                <button type="button" class="close" data-dismiss="modal">&times;</button>
+                <h4 class="modal-title">Add An Agency</h4>
+            </div>
+            <div class="modal-body">
+                <form  class="form-horizontal" role="form" >
+
+                    <div class="form-group">
+                        <label for="firstName" class="col-sm-3 control-label">Name</label>
+                        <div class="col-sm-9">
+                            <input type="text" id="firstName" name="Fname" placeholder="First Name" class="form-control" autofocus>
+
+                        </div>
+                    </div>
+                    <div class="form-group">
+                        <label for="firstName" class="col-sm-3 control-label">description</label>
+                        <div class="col-sm-9">
+                            <input type="text" id="lastName" name="Lname" placeholder="About" class="form-control" autofocus>
+
+                        </div>
+                    </div>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-default" data-dismiss="modal">Submit</button>
+            </div>
+        </div>
+
+    </div>
+</div>
