+++ views/login.ejs
@@ -1,16 +1,27 @@
 <!DOCTYPE html>
-<%include Header%>
+<link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css">
+<%include Header %>
+<div class="container">
 
-<body>
-    <form method="post" action="/login">
-            <label>Username:</label>
-            <input type="text" name="username">
+    <form method ="post" class="form-horizontal" role="form"  action="/login">
+        <div class="form-group">
+            <label for="username" class="col-sm-3 control-label">User Name</label>
+            <div class="col-sm-9">
+                <input type="text" id="username" name="username" placeholder="Email" class="form-control" autofocus>
 
-            <label>Password:</label>
-            <input type="password" name="password">
+            </div>
+        </div>
+        <div class="form-group">
+            <label for="password" class="col-sm-3 control-label">Password</label>
+            <div class="col-sm-9">
+                <input type="password" id="password" name="password" placeholder="password" class="form-control" autofocus>
 
-            <input type="submit" value="Login">
-
-    </form>
-</body>
-
+            </div>
+        </div>
+        <div class="form-group">
+            <div class="col-sm-9 col-sm-offset-3">
+                <button type="submit" class="btn btn-primary btn-block">Login</button>
+            </div>
+        </div>
+    </form> <!-- /form -->
+</div> <!-- ./container -->
